backend/requirements.txt
Django==4.2.10
whitenoise==6.6.0
gunicorn==21.2.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
